requirements.txt
streamlit
python-chess
streamlit-chessboard
google-genai
